Move Actions Matrix Expansion to Sdk
* removes copy in codelens
